Updated master to v10

+++ org.adempiere.base/src/org/compiere/model/I_A_Asset.java
@@ -22,7 +22,7 @@ import org.compiere.util.KeyNamePair;
 
 /** Generated Interface for A_Asset
  *  @author iDempiere (generated) 
- *  @version Release 9
+ *  @version Release 10
  */
 public interface I_A_Asset 
 {
@@ -148,8 +148,8 @@ public interface I_A_Asset
     /** Column name AD_Client_ID */
     public static final String COLUMNNAME_AD_Client_ID = "AD_Client_ID";
 
-	/** Get Client.
-	  * Client/Tenant for this installation.
+	/** Get Tenant.
+	  * Tenant for this installation.
 	  */
 	public int getAD_Client_ID();
 
@@ -157,12 +157,12 @@ public interface I_A_Asset
     public static final String COLUMNNAME_AD_Org_ID = "AD_Org_ID";
 
 	/** Set Organization.
-	  * Organizational entity within client
+	  * Organizational entity within tenant
 	  */
 	public void setAD_Org_ID (int AD_Org_ID);
 
 	/** Get Organization.
-	  * Organizational entity within client
+	  * Organizational entity within tenant
 	  */
 	public int getAD_Org_ID();
 
@@ -694,13 +694,13 @@ public interface I_A_Asset
     /** Column name NextMaintenenceDate */
     public static final String COLUMNNAME_NextMaintenenceDate = "NextMaintenenceDate";
 
-	/** Set Next Maintenence.
-	  * Next Maintenence Date
+	/** Set Next Maintenance.
+	  * Next Maintenance Date
 	  */
 	public void setNextMaintenenceDate (Timestamp NextMaintenenceDate);
 
-	/** Get Next Maintenence.
-	  * Next Maintenence Date
+	/** Get Next Maintenance.
+	  * Next Maintenance Date
 	  */
 	public Timestamp getNextMaintenenceDate();
 
@@ -800,12 +800,12 @@ public interface I_A_Asset
 	/** Set Usable Life - Years.
 	  * Years of the usable life of the asset
 	  */
-	public void setUseLifeYears (int UseLifeYears);
+	public void setUseLifeYears (BigDecimal UseLifeYears);
 
 	/** Get Usable Life - Years.
 	  * Years of the usable life of the asset
 	  */
-	public int getUseLifeYears();
+	public BigDecimal getUseLifeYears();
 
     /** Column name UseUnits */
     public static final String COLUMNNAME_UseUnits = "UseUnits";
