+++ org.adempiere.base/src/org/compiere/model/I_A_Asset_Change.java
@@ -22,7 +22,7 @@ import org.compiere.util.KeyNamePair;
 
 /** Generated Interface for A_Asset_Change
  *  @author iDempiere (generated) 
- *  @version Release 9
+ *  @version Release 10
  */
 public interface I_A_Asset_Change 
 {
@@ -148,8 +148,8 @@ public interface I_A_Asset_Change
     /** Column name AD_Client_ID */
     public static final String COLUMNNAME_AD_Client_ID = "AD_Client_ID";
 
-	/** Get Client.
-	  * Client/Tenant for this installation.
+	/** Get Tenant.
+	  * Tenant for this installation.
 	  */
 	public int getAD_Client_ID();
 
@@ -228,12 +228,12 @@ public interface I_A_Asset_Change
     public static final String COLUMNNAME_AD_Org_ID = "AD_Org_ID";
 
 	/** Set Organization.
-	  * Organizational entity within client
+	  * Organizational entity within tenant
 	  */
 	public void setAD_Org_ID (int AD_Org_ID);
 
 	/** Get Organization.
-	  * Organizational entity within client
+	  * Organizational entity within tenant
 	  */
 	public int getAD_Org_ID();
 
@@ -419,19 +419,19 @@ public interface I_A_Asset_Change
     /** Column name AssetAccumDepreciationAmt */
     public static final String COLUMNNAME_AssetAccumDepreciationAmt = "AssetAccumDepreciationAmt";
 
-	/** Set AssetAccumDepreciationAmt	  */
+	/** Set Asset Accum Depreciation Amt	  */
 	public void setAssetAccumDepreciationAmt (BigDecimal AssetAccumDepreciationAmt);
 
-	/** Get AssetAccumDepreciationAmt	  */
+	/** Get Asset Accum Depreciation Amt	  */
 	public BigDecimal getAssetAccumDepreciationAmt();
 
     /** Column name AssetBookValueAmt */
     public static final String COLUMNNAME_AssetBookValueAmt = "AssetBookValueAmt";
 
-	/** Set AssetBookValueAmt	  */
+	/** Set Asset Book Value Amt	  */
 	public void setAssetBookValueAmt (BigDecimal AssetBookValueAmt);
 
-	/** Get AssetBookValueAmt	  */
+	/** Get Asset Book Value Amt	  */
 	public BigDecimal getAssetBookValueAmt();
 
     /** Column name AssetDepreciationDate */
@@ -547,28 +547,28 @@ public interface I_A_Asset_Change
     /** Column name ChangeAmt */
     public static final String COLUMNNAME_ChangeAmt = "ChangeAmt";
 
-	/** Set ChangeAmt	  */
+	/** Set Change Amt	  */
 	public void setChangeAmt (BigDecimal ChangeAmt);
 
-	/** Get ChangeAmt	  */
+	/** Get Change Amt	  */
 	public BigDecimal getChangeAmt();
 
     /** Column name ChangeDate */
     public static final String COLUMNNAME_ChangeDate = "ChangeDate";
 
-	/** Set ChangeDate	  */
+	/** Set Change Date	  */
 	public void setChangeDate (Timestamp ChangeDate);
 
-	/** Get ChangeDate	  */
+	/** Get Change Date	  */
 	public Timestamp getChangeDate();
 
     /** Column name ChangeType */
     public static final String COLUMNNAME_ChangeType = "ChangeType";
 
-	/** Set ChangeType	  */
+	/** Set Change Type	  */
 	public void setChangeType (String ChangeType);
 
-	/** Get ChangeType	  */
+	/** Get Change Type	  */
 	public String getChangeType();
 
     /** Column name C_Location_ID */
@@ -737,12 +737,12 @@ public interface I_A_Asset_Change
     /** Column name PostingType */
     public static final String COLUMNNAME_PostingType = "PostingType";
 
-	/** Set PostingType.
+	/** Set Posting Type.
 	  * The type of posted amount for the transaction
 	  */
 	public void setPostingType (String PostingType);
 
-	/** Get PostingType.
+	/** Get Posting Type.
 	  * The type of posted amount for the transaction
 	  */
 	public String getPostingType();
@@ -804,12 +804,12 @@ public interface I_A_Asset_Change
 	/** Set Usable Life - Years.
 	  * Years of the usable life of the asset
 	  */
-	public void setUseLifeYears (int UseLifeYears);
+	public void setUseLifeYears (BigDecimal UseLifeYears);
 
 	/** Get Usable Life - Years.
 	  * Years of the usable life of the asset
 	  */
-	public int getUseLifeYears();
+	public BigDecimal getUseLifeYears();
 
     /** Column name UseUnits */
     public static final String COLUMNNAME_UseUnits = "UseUnits";
